--- sktime/distances/adapters/__init__.py
@@ -1,6 +1,0 @@
-"""Module containing adapters for distance related estimators."""
-# copyright: sktime developers, BSD-3-Clause License (see LICENSE file)
-
-__all__ = ["KNeighborsSklearnAdapter"]
-
-from sktime.distances.adapters.sklearn import KNeighborsSklearnAdapter
