+++ sktime/dists_kernels/base/adapters/__init__.py
@@ -1,5 +1,6 @@
 """Adapter mixins for dists_kernels module."""
 
+from sktime.dists_kernels.base.adapters._sklearn import _SklearnKnnAdapter
 from sktime.dists_kernels.base.adapters._tslearn import _TslearnPwTrafoAdapter
 
-__all__ = ["_TslearnPwTrafoAdapter"]
+__all__ = ["_TslearnPwTrafoAdapter", "_SklearnKnnAdapter"]
