+++ sktime/dists_kernels/base/adapters/_sklearn.py
@@ -1,7 +1,8 @@
-"""Implements adapter for sklearn distance related estimators."""
+# copyright: sktime developers, BSD-3-Clause License (see LICENSE file)
+"""Implements adapter for sklearn distances."""
 
-__author__ = ["Z-Fran"]
-__all__ = ["KNeighborsSklearnAdapter"]
+__author__ = ["fkiraly", "Z-Fran"]
+__all__ = ["_SklearnKnnAdapter"]
 
 import numpy as np
 import pandas as pd
@@ -9,13 +10,13 @@ import pandas as pd
 from sktime.datatypes import convert
 
 
-class KNeighborsSklearnAdapter:
-    """Mixin adapter class for sklearn KNeighbors models."""
+class _SklearnKnnAdapter:
+    """Base adapter mixin for sklearn distances of KNeighbors."""
 
     _tags = {
         # packaging info
         # --------------
-        "authors": ["Z-Fran"],
+        "authors": ["fkiraly", "Z-Fran"],
         # estimator type
         # --------------
         "capability:multivariate": True,
@@ -83,87 +84,3 @@ class KNeighborsSklearnAdapter:
         else:
             X_mtype = "pd-multiindex"
         return convert(X, from_type=X_mtype, to_type="numpyflat")
-
-    def _fit(self, X, y):
-        """Fit the model using X as training data and y as target values.
-
-        Parameters
-        ----------
-        X : sktime compatible Panel data container, of mtype X_inner_mtype,
-            with n time series to fit the estimator to
-        y : {array-like, sparse matrix}
-            Target values of shape = [n]
-        """
-        self.n_vars_ = X.shape[1]
-        if self.algorithm == "brute":
-            return self._fit_precomp(X=X, y=y)
-        else:
-            return self._fit_dist(X=X, y=y)
-
-    def kneighbors(self, X, n_neighbors=None, return_distance=True):
-        """Find the K-neighbors of a point.
-
-        Returns indices of and distances to the neighbors of each point.
-
-        Parameters
-        ----------
-        X : sktime-compatible data format, Panel or Series, with n_samples series
-        n_neighbors : int
-            Number of neighbors to get (default is the value
-            passed to the constructor).
-        return_distance : boolean, optional. Defaults to True.
-            If False, distances will not be returned
-
-        Returns
-        -------
-        dist : array
-            Array representing the lengths to points, only present if
-            return_distance=True
-        ind : array
-            Indices of the nearest points in the population matrix.
-        """
-        self.check_is_fitted()
-
-        # boilerplate input checks for predict-like methods
-        X = self._check_convert_X_for_predict(X)
-
-        # self._X should be the stored _X
-        dist_mat = self._distance(X, self._X)
-
-        result = self.knn_estimator_.kneighbors(
-            dist_mat, n_neighbors=n_neighbors, return_distance=return_distance
-        )
-
-        # result is either dist, or (dist, ind) pair, depending on return_distance
-        return result
-
-    def _predict(self, X):
-        """Predict the class labels for the provided data.
-
-        Parameters
-        ----------
-        X : sktime-compatible Panel data, of mtype X_inner_mtype, with n_samples series
-            data to predict class labels for
-
-        Returns
-        -------
-        y : array of shape [n_samples] or [n_samples, n_outputs]
-            Class labels for each data sample.
-        """
-        if self.algorithm == "brute":
-            return self._predict_precomp(X)
-        else:
-            return self._predict_dist(X)
-
-    def _predict_dist(self, X):
-        """Predict using adapted distance metric."""
-        X = self._convert_X_to_sklearn(X)
-        y_pred = self.knn_estimator_.predict(X)
-        return y_pred
-
-    def _predict_precomp(self, X):
-        """Predict using precomputed distance matrix."""
-        # self._X should be the stored _X
-        dist_mat = self._distance(X, self._X)
-        y_pred = self.knn_estimator_.predict(dist_mat)
-        return y_pred
